+++ com.revolsys.open.core/src/main/java/com/revolsys/elevation/gridded/GriddedElevationModelHillshadeDataBuffer.java
@@ -31,7 +31,7 @@ public class GriddedElevationModelHillshadeDataBuffer extends DataBuffer {
     this.elevationModel = elevationModel;
     this.width = elevationModel.getGridWidth();
     this.height = elevationModel.getGridHeight();
-    final int cellSize = elevationModel.getGridCellSize();
+    final double cellSize = elevationModel.getGridCellSize();
     this.oneDivCellSizeTimes8 = 1.0 / (8 * cellSize);
 
     setZenithDegrees(45.0);
